fixed image typing

+++ nnes/utils.py
@@ -16,4 +16,4 @@ def get_tensors_from_image(image: Tensor) -> Tuple[Tensor, Tensor]:
     return X, Y
 
 def get_image_from_tensor(Y: Tensor, image_shape: Tuple) -> Tensor:
-    return 255 * Y.view(*image_shape)
+    return 255 * Y.view(*image_shape).to(torch.uint8)
